# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sensible",
        "title": "Sensible Query Spec",
        "contact": {
            "name": "sensiblequery",
            "url": "https://github.com/sensible-contract/sensiblequery",
            "email": "jiedohh@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/nft/auction/utxo-detail/{codehash}/{nftid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UTXO",
                    "token NFT"
                ],
                "summary": "通过拍卖的CodeHash和NFT ID获取具体NFTAuction合约utxo",
                "parameters": [
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Auction Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "NFT ID",
                        "name": "nftid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "仅返回ready状态的记录",
                        "name": "ready",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.NFTAuctionResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/nft/sell/utxo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UTXO",
                    "token NFT"
                ],
                "summary": "获取NFTSell合约的utxo列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.NFTSellResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/nft/sell/utxo-by-address/{address}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UTXO",
                    "token NFT"
                ],
                "summary": "通过出售人地址获取NFTSell合约utxo列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "17SkEw2md5avVNyYgj6RiXuQKNwkXaxFyQ",
                        "description": "Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.NFTSellResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/nft/sell/utxo-detail/{codehash}/{genesis}/{token_index}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UTXO",
                    "token NFT"
                ],
                "summary": "通过NFT的CodeHash+溯源genesis和NFT Token Index获取具体NFTSell合约utxo",
                "parameters": [
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 3,
                        "description": "Token Index",
                        "name": "token_index",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": true,
                        "description": "仅返回ready状态的记录",
                        "name": "ready",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.NFTSellResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/nft/sell/utxo/{codehash}/{genesis}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UTXO",
                    "token NFT"
                ],
                "summary": "通过NFT的CodeHash+溯源genesis获取NFTSell合约utxo列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始游标",
                        "name": "cursor",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "返回记录数量",
                        "name": "size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "844c56bb99afc374967a27ce3b46244e2e1fba60",
                        "description": "Code Hash160",
                        "name": "codehash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "74967a27ce3b46244e2e1fba60844c56bb99afc3",
                        "description": "Genesis ID",
                        "name": "genesis",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": [{}], \"msg\": \"ok\"}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.NFTSellResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.NFTAuctionResp": {
            "type": "object",
            "properties": {
                "bidBsvPrice": {
                    "description": "当前拍卖NFT的出价价格(satoshi)",
                    "type": "integer"
                },
                "bidTimestamp": {
                    "description": "当前拍卖出价的时间戳",
                    "type": "integer"
                },
                "bidderAddress": {
                    "description": "当前拍卖出价人的地址",
                    "type": "string"
                },
                "codehash": {
                    "description": "当前拍卖NFT合约hash160(CodePart)",
                    "type": "string"
                },
                "feeAddress": {
                    "description": "当前拍卖手续费的地址",
                    "type": "string"
                },
                "feeAmount": {
                    "description": "当前拍卖手续费",
                    "type": "integer"
                },
                "feeRate": {
                    "description": "当前拍卖手续费比例",
                    "type": "integer"
                },
                "genesis": {
                    "description": "当前拍卖NFT的genesis",
                    "type": "string"
                },
                "height": {
                    "description": "当前交易被打包的区块高度",
                    "type": "integer"
                },
                "idx": {
                    "description": "输出被花费的txid所在区块内序号",
                    "type": "integer"
                },
                "isReady": {
                    "description": "当前拍卖NFT是否已准备好(转出到拍卖合约)",
                    "type": "boolean"
                },
                "nftCodehash": {
                    "description": "当前拍卖NFT的codehash",
                    "type": "string"
                },
                "nftID": {
                    "description": "当前拍卖NFT的nftId",
                    "type": "string"
                },
                "satoshi": {
                    "description": "拍卖合约输出的satoshi",
                    "type": "integer"
                },
                "senderAddress": {
                    "description": "当前拍卖发起人的地址",
                    "type": "string"
                },
                "sensibleId": {
                    "description": "当前拍卖NFT合约的sensibleId，即genesisTx的outpoint，Hex编码",
                    "type": "string"
                },
                "startBsvPrice": {
                    "description": "当前拍卖NFT的起拍价格(satoshi)",
                    "type": "integer"
                },
                "txid": {
                    "description": "拍卖合约txid",
                    "type": "string"
                },
                "vout": {
                    "description": "拍卖合约输出序号",
                    "type": "integer"
                }
            }
        },
...
        "model.NFTSellResp": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "当前售卖人seller的address",
                    "type": "string"
                },
                "codehash": {
                    "description": "当前售卖NFT合约hash160(CodePart)",
                    "type": "string"
                },
                "genesis": {
                    "description": "当前售卖NFT的genesis",
                    "type": "string"
                },
                "height": {
                    "description": "当前交易被打包的区块高度",
                    "type": "integer"
                },
                "idx": {
                    "description": "输出被花费的txid所在区块内序号",
                    "type": "integer"
                },
                "isReady": {
                    "description": "当前售卖NFT是否已准备好(转出到售卖合约)",
                    "type": "boolean"
                },
                "metaOutputIndex": {
                    "description": "当前售卖NFT的metaOutputIndex",
                    "type": "integer"
                },
                "metaTxId": {
                    "description": "当前售卖NFT的metaTxId",
                    "type": "string"
                },
                "price": {
                    "description": "当前售卖NFT的出售价格(satoshi)",
                    "type": "integer"
                },
                "satoshi": {
                    "description": "售卖合约输出的satoshi",
                    "type": "integer"
                },
                "sensibleId": {
                    "description": "当前售卖NFT合约的sensibleId，即genesisTx的outpoint，Hex编码",
                    "type": "string"
                },
                "supply": {
                    "description": "当前售卖NFT最大发行量",
                    "type": "integer"
                },
                "tokenIndex": {
                    "description": "当前售卖NFT的tokenIndex",
                    "type": "string"
                },
                "txid": {
                    "description": "售卖合约txid",
                    "type": "string"
                },
                "vout": {
                    "description": "售卖合约输出序号",
                    "type": "integer"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}